Refactor solochain template
Refactor solochain template
