+++ templates/solochain/node/Cargo.toml
@@ -51,7 +51,8 @@ sp-block-builder = { workspace = true, default-features = true }
 
 # frame and pallets
 frame-system = { workspace = true, default-features = true }
-pallet-transaction-payment = { workspace = true }
+frame-metadata-hash-extension = { workspace = true, default-features = true }
+pallet-transaction-payment = { workspace = true, default-features = true }
 pallet-transaction-payment-rpc = { workspace = true, default-features = true }
 substrate-frame-rpc-system = { workspace = true, default-features = true }
